move dynamic class injection integration tests to their own module
